switch timestamps to f64 and performance.now()
- f64 is the best way to represent a JS number in Rust, as it doesn't
  require manual conversion like u64 and doesn't silently overflow like
  u32 does,
- performance.now() makes a bit mor sense to use, although
  Date.now() was fine so far. But I rather make the change now, since
  the extra resolution might be needed in the future. And technically,
  Date.now() isn't monotonic, so it's also just cleaner to use
  performance.now().

+++ bouncy_instructor/src/intern/step/detection.rs
@@ -7,16 +7,16 @@ impl DetectedStep {
     pub(crate) fn new(step_name: String, poses: Vec<PoseApproximation>) -> Self {
         Self {
             step_name,
-            start: poses.first().map(|p| p.timestamp).unwrap_or(0),
-            end: poses.last().map(|p| p.timestamp).unwrap_or(0),
+            start: poses.first().map(|p| p.timestamp).unwrap_or(0.0),
+            end: poses.last().map(|p| p.timestamp).unwrap_or(0.0),
             error: poses.iter().map(|p| p.error).sum::<f32>() / poses.len() as f32,
             poses,
         }
     }
 
     pub(crate) fn update_stats(&mut self) {
-        self.start = self.poses.first().map(|p| p.timestamp).unwrap_or(0);
-        self.end = self.poses.last().map(|p| p.timestamp).unwrap_or(0);
+        self.start = self.poses.first().map(|p| p.timestamp).unwrap_or(0.0);
+        self.end = self.poses.last().map(|p| p.timestamp).unwrap_or(0.0);
         self.error = self.poses.iter().map(|p| p.error).sum::<f32>() / self.poses.len() as f32;
     }
 }
@@ -52,8 +52,8 @@ impl Tracker {
         }
         let end = end.min(self.timestamps.len());
         let dt = 60_000.0 / self.detector.bpm;
-        let min_dt = (dt * 0.5).round() as u64;
-        let max_dt = (dt * 1.5).round() as u64;
+        let min_dt = (dt * 0.5).round() as f64;
+        let max_dt = (dt * 1.5).round() as f64;
 
         let threshold = 0.2;
         let pose_window_ms = max_dt;
@@ -95,7 +95,7 @@ impl Tracker {
                 // satisfies v0
                 start = step_start_index + 1;
                 // recover i0 and i1
-                start_t = *self.timestamps.get(start).unwrap_or(&u64::MAX);
+                start_t = *self.timestamps.get(start).unwrap_or(&f64::MAX);
             } else {
                 // couldn't even match a pose, shift search window by one window length
                 start_t += pose_window_ms;
@@ -126,8 +126,8 @@ impl Tracker {
     fn detect_step(
         &self,
         start: usize,
-        min_dt: u64,
-        max_dt: u64,
+        min_dt: f64,
+        max_dt: f64,
         db: &DanceCollection,
     ) -> Option<DetectedStep> {
         let mut best_error = f32::INFINITY;
@@ -181,7 +181,7 @@ mod tests {
     #[test]
     fn test_detect_dance_front() {
         let degrees = [0, 0, 45, 90];
-        let times = [0, 100, 500, 1000];
+        let times = [0., 100., 500., 1000.];
         let expected_steps = ["Test-Step-2"];
 
         check_detect_front_dance(&degrees, times, &expected_steps);
@@ -190,7 +190,7 @@ mod tests {
     #[test]
     fn test_detect_dance_side_no_match() {
         let degrees = [0, 0, 45, 90];
-        let times = [0, 100, 500, 1000];
+        let times = [0., 100., 500., 1000.];
         let expected_steps = [];
 
         check_detect_side_dance(&degrees, times, &expected_steps);
@@ -199,7 +199,7 @@ mod tests {
     #[test]
     fn test_detect_dance_side() {
         let degrees = [0, 0, 90, 45, 45, 90];
-        let times = [0, 100, 1000, 2000, 3000, 4000];
+        let times = [0., 100., 1000., 2000., 3000., 4000.];
         let expected_steps = ["Test-Step-4"];
 
         check_detect_side_dance(&degrees, times, &expected_steps);
@@ -209,7 +209,7 @@ mod tests {
     #[test]
     fn test_detect_dance_skip_start() {
         let degrees = [-45, -45, 0, 90];
-        let times = [0, 100, 500, 1500];
+        let times = [0., 100., 500., 1500.];
         let expected_steps = ["Test-Step-2"];
 
         check_detect_front_dance(&degrees, times, &expected_steps);
@@ -218,7 +218,7 @@ mod tests {
     #[test]
     fn test_detect_dance_multi_step() {
         let degrees = [0, -90, 0, -90, -90, 90, 0, -90, 0, -90, 0, 90];
-        let times = (0..degrees.len()).map(|n| (n * 1000) as u64);
+        let times = (0..degrees.len()).map(|n| (n * 1000) as f64);
         let expected_steps = [
             "Test-Step-1", // 0,-90
             "Test-Step-1", // 0,-90
@@ -238,7 +238,7 @@ mod tests {
             -90, -45, -45, -45, 0, 0, 0, 0, 0, 0, 0, // 0 - 900ms
             0, 45, 45, 90, 90, 90, 90, 90, 90, 90, // 1000 - 1900ms
         ];
-        let times = (0..degrees.len()).map(|n| (n * 100) as u64);
+        let times = (0..degrees.len()).map(|n| (n * 100) as f64);
         let expected_steps = ["Test-Step-2"];
 
         check_detect_front_dance(&degrees, times, &expected_steps);
@@ -247,7 +247,7 @@ mod tests {
     #[track_caller]
     fn check_detect_side_dance(
         degrees: &[i16],
-        times: impl IntoIterator<Item = u64>,
+        times: impl IntoIterator<Item = f64>,
         expected_steps: &[&str],
     ) {
         let kp = facing_right_keypoints();
@@ -256,7 +256,7 @@ mod tests {
     #[track_caller]
     fn check_detect_front_dance(
         degrees: &[i16],
-        times: impl IntoIterator<Item = u64>,
+        times: impl IntoIterator<Item = f64>,
         expected_steps: &[&str],
     ) {
         let kp = facing_camera_keypoints();
@@ -267,7 +267,7 @@ mod tests {
     fn check_detect_dance(
         mut kp: Keypoints,
         degrees: &[i16],
-        times: impl Iterator<Item = u64>,
+        times: impl Iterator<Item = f64>,
         expected_steps: &[&str],
     ) {
         setup();
